Dissect tf_display into several classes
